requirement.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
openai==1.43.0
